Move aggregate functions into own package
