+++ h2/src/main/org/h2/expression/aggregate/AggregateDataDefault.java
@@ -3,10 +3,10 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
-package org.h2.expression;
+package org.h2.expression.aggregate;
 
 import org.h2.engine.Database;
-import org.h2.expression.Aggregate.AggregateType;
+import org.h2.expression.aggregate.Aggregate.AggregateType;
 import org.h2.message.DbException;
 import org.h2.util.ValueHashMap;
 import org.h2.value.DataType;
