+++ h2/src/main/org/h2/expression/aggregate/AggregateDataMedian.java
@@ -3,7 +3,7 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
-package org.h2.expression;
+package org.h2.expression.aggregate;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -14,6 +14,8 @@ import org.h2.engine.Database;
 import org.h2.engine.Mode;
 import org.h2.engine.Session;
 import org.h2.engine.SysProperties;
+import org.h2.expression.Expression;
+import org.h2.expression.ExpressionColumn;
 import org.h2.index.Cursor;
 import org.h2.index.Index;
 import org.h2.result.SearchRow;
